Calendar runs in container
Prior to now there was not a way to pick dates because there was no
other UI elemnts that allowed that. Now there's room for them, but it is
not currently set for now.
